requirements.txt
fastapi[standard]
audio-similarity==1.0.0
sqlalchemy
passlib[bcrypt]
python-jose
python-multipart
psycopg2-binary
python-dotenv

numpy~=2.0.2
tensorflow~=2.18.0
soundfile~=0.13.0
librosa~=0.10.2.post1
scikit-learn~=1.6.1
keras~=3.8.0